--- src/components/blocks/Form/BaseFormTextField.tsx
@@ -1,59 +1,0 @@
-import type { PropsWithChildren } from 'react'
-
-import { useFieldContext } from '@hooks/features/forms/useForm'
-import { FieldError, TextField, type TextFieldProps } from '@ui/Form/Form'
-import { Label } from '@ui/Typography/Text'
-import type { CommonFormFieldProps } from '@blocks/Form/types'
-
-export type BaseFormTextFieldProps = CommonFormFieldProps & {
-  inputMode?: TextFieldProps['inputMode']
-  isTextArea?: boolean
-  errorText?: string
-}
-
-export function BaseFormTextField({
-  label,
-  inline = false,
-  showLabel = true,
-  showFieldError = true,
-  isTextArea = false,
-  isReadOnly = false,
-  errorText,
-  className,
-  children,
-}: PropsWithChildren<BaseFormTextFieldProps>) {
-  const field = useFieldContext<string>()
-
-  const { name, state } = field
-  const { meta } = state
-  const { errors, isValid } = meta
-
-  const errorMessage = errors.length !== 0 ? (errors[0] as string) : errorText
-  const shouldShowErrorMessage = showFieldError && errorMessage
-
-  const additionalAriaProps = !showLabel && { 'aria-label': label }
-  return (
-    <TextField
-      name={name}
-      isInvalid={!isValid || !!errorText}
-      inline={inline}
-      className={className}
-      textarea={isTextArea}
-      isReadOnly={isReadOnly}
-      {...additionalAriaProps}
-    >
-      {showLabel && (
-        <Label
-          slot='label'
-          size='sm'
-          htmlFor={name}
-          {...(!inline && { pbe: '3xs' })}
-        >
-          {label}
-        </Label>
-      )}
-      {children}
-      {shouldShowErrorMessage && <FieldError>{errorMessage}</FieldError>}
-    </TextField>
-  )
-}
