--- src/components/blocks/Form/FormSwitchField.tsx
@@ -1,58 +1,0 @@
-import { type PropsWithChildren, type ReactNode } from 'react'
-import classNames from 'classnames'
-
-import { useFieldContext } from '@hooks/features/forms/useForm'
-import { HStack } from '@ui/Stack/Stack'
-import { Switch } from '@ui/Switch/Switch'
-import { Label } from '@ui/Typography/Text'
-import type { CommonFormFieldProps } from '@blocks/Form/types'
-
-import './formSwitchField.scss'
-
-export type FormSwitchFieldProps = CommonFormFieldProps & {
-  slots?: {
-    LabelIcon?: ReactNode
-  }
-}
-const FORM_SWITCH_FIELD_CLASSNAME = 'Layer__FormSwitchField'
-export function FormSwitchField({
-  label,
-  slots,
-  className,
-  inline = false,
-  showLabel = true,
-  isReadOnly = false,
-}: PropsWithChildren<FormSwitchFieldProps>) {
-  const field = useFieldContext<boolean>()
-
-  const { name, state, handleChange } = field
-  const { value } = state
-
-  const additionalAriaProps = !showLabel && { 'aria-label': label }
-
-  const switchClassNames = classNames(
-    FORM_SWITCH_FIELD_CLASSNAME,
-    inline && `${FORM_SWITCH_FIELD_CLASSNAME}--inline`,
-    className,
-  )
-
-  return (
-    <Switch
-      className={switchClassNames}
-      isSelected={value}
-      onChange={handleChange}
-      name={name}
-      isReadOnly={isReadOnly}
-      slot='switch'
-      {...additionalAriaProps}
-    >
-      {showLabel
-        && (
-          <HStack slot='label' gap='xs' align='center'>
-            {slots?.LabelIcon}
-            <Label size='sm' htmlFor={name}>{label}</Label>
-          </HStack>
-        )}
-    </Switch>
-  )
-}
