--- src/components/blocks/Form/formDatePickerField.scss
@@ -1,25 +1,0 @@
-.Layer__FormDatePickerField {
-  display: grid;
-
-  grid-template-areas:
-    'label'
-    'datepicker';
-
-  &--inline {
-    grid-template-areas: 'label datepicker';
-
-    align-items: center;
-
-    .Layer__UI__DatePicker {
-      display: contents;
-
-      > .Layer__UI__Label {
-        grid-area: label;
-      }
-    }
-  }
-
-  .Layer__UI__DatePicker {
-    grid-area: datepicker;
-  }
-}
