--- src/components/blocks/Form/formRadioGroupField.scss
@@ -1,22 +1,0 @@
-.Layer__FormRadioGroupField {
-  &.Layer__RadioGroup {
-    display: grid;
-    grid-template-areas:
-      'label'
-      'options';
-    gap: var(--spacing-3xs);
-
-    &.Layer__FormRadioGroupField--inline {
-      grid-template-areas: 'label options';
-      gap: 0;
-    }
-
-    > [slot='label'] {
-      grid-area: label;
-    }
-
-    > [slot='options'] {
-      grid-area: options;
-    }
-  }
-}
