+++ src/components/blocks/forms/FormSwitchField.tsx
@@ -1,0 +1,34 @@
+import { type ReactNode } from 'react'
+
+import { useFieldContext } from '@hooks/features/forms/useForm'
+import { Switch } from '@ui/Switch/Switch'
+import { FormFieldShell, useFormField } from '@blocks/forms/FormFieldShell'
+import type { CommonFormFieldProps } from '@blocks/forms/types'
+
+export type FormSwitchFieldProps = CommonFormFieldProps & {
+  slots?: {
+    LabelIcon?: ReactNode
+  }
+}
+
+export function FormSwitchField({ slots, ...props }: FormSwitchFieldProps) {
+  const field = useFieldContext<boolean>()
+
+  const { state, handleChange, handleBlur } = field
+  const { value } = state
+
+  const { name, rootProps, shellProps } = useFormField(props)
+
+  return (
+    <Switch
+      {...rootProps}
+      isSelected={value}
+      onChange={handleChange}
+      onBlur={handleBlur}
+      name={name}
+    >
+      {/* Switch has no invalid state in react-aria, so only the error text reflects validation. */}
+      <FormFieldShell {...shellProps} slots={slots} />
+    </Switch>
+  )
+}
