+++ src/components/blocks/forms/FormTextAreaField.tsx
@@ -2,7 +2,7 @@ import { useCallback } from 'react'
 
 import { useFieldContext } from '@hooks/features/forms/useForm'
 import { TextArea } from '@ui/Input/TextArea'
-import { BaseFormTextField, type BaseFormTextFieldProps } from '@blocks/Form/BaseFormTextField'
+import { BaseFormTextField, type BaseFormTextFieldProps } from '@blocks/forms/BaseFormTextField'
 
 export type FormTextAreaFieldProps = Omit<BaseFormTextFieldProps, 'isTextArea'> & {
   placeholder?: string
