+++ src/components/blocks/forms/_formFieldMixins.scss
@@ -1,0 +1,76 @@
+/*
+ * Layout and state building blocks shared by every field in `@blocks/forms`. Fields differ only in
+ * which slot holds their control, so all control slots are matched as one group.
+ */
+$control-slots: "[slot='input'], [slot='checkbox'], [slot='indicator'], [slot='options']";
+
+@mixin controls {
+  > :is(#{$control-slots}) {
+    @content;
+  }
+}
+
+@mixin stacked-layout {
+  display: grid;
+  grid-template-areas:
+    'label'
+    'control'
+    'error';
+  grid-template-columns: minmax(0, 1fr);
+
+  > [slot='label'] {
+    grid-area: label;
+
+    padding-block-end: var(--spacing-3xs);
+  }
+
+  @include controls {
+    grid-area: control;
+  }
+
+  > [slot='errorMessage'] {
+    grid-area: error;
+  }
+}
+
+@mixin inline-layout {
+  grid-template-areas:
+    'label control'
+    '. error';
+  grid-template-columns: var(--form-field-label-inline-size, auto) minmax(0, 1fr);
+  column-gap: var(--spacing-xs);
+
+  align-items: center;
+
+  > [slot='label'] {
+    padding-block-end: 0;
+  }
+
+  &[data-textarea] {
+    align-items: flex-start;
+  }
+}
+
+@mixin centered-control {
+  justify-content: center;
+
+  @include controls {
+    justify-self: center;
+
+    margin-block: var(--spacing-2xs);
+  }
+}
+
+@mixin field-states {
+  &[data-readonly] > [slot='input'] {
+    background-color: var(--bg-muted);
+  }
+
+  &[data-disabled] {
+    cursor: not-allowed;
+
+    > [slot='label'] {
+      color: var(--color-base-500);
+    }
+  }
+}
