+++ src/components/blocks/forms/types.ts
@@ -2,7 +2,10 @@ export interface CommonFormFieldProps {
   label: string
   className?: string
   inline?: boolean
+  align?: 'start' | 'center'
   showLabel?: boolean
   showFieldError?: boolean
   isReadOnly?: boolean
+  isDisabled?: boolean
+  errorText?: string
 }
