+++ src/components/features/bankTransactions/RecordBankTransactionForm/RecordBankTransactionForm.tsx
@@ -15,7 +15,7 @@ import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { useSizeClass } from '@hooks/utils/size/useWindowSize'
 import { Form } from '@ui/Form/Form'
 import { Label } from '@ui/Typography/Text'
-import { FieldErrors } from '@blocks/Form/FieldErrors'
+import { FieldErrors } from '@blocks/forms/FieldErrors'
 import { BankTransactionTaxCodeSelect } from '@features/bankTransactions/BankTransactionTaxCodeSelect/BankTransactionTaxCodeSelect'
 import { RecordBankTransactionFormCategoryComboBox } from '@features/bankTransactions/RecordBankTransactionForm/RecordBankTransactionFormCategoryComboBox'
 import { type RecordBankTransactionFormApi, type RecordBankTransactionVariant } from '@features/bankTransactions/RecordBankTransactionForm/useRecordBankTransactionForm'
