+++ src/components/features/mileage/TripForm/TripForm.tsx
@@ -11,7 +11,7 @@ import { Button } from '@ui/Button/Button'
 import { DataState, DataStateStatus } from '@ui/DataState/DataState'
 import { Form } from '@ui/Form/Form'
 import { HStack, VStack } from '@ui/Stack/Stack'
-import { FieldErrors } from '@blocks/Form/FieldErrors'
+import { FieldErrors } from '@blocks/forms/FieldErrors'
 import { TripAddressComboBox } from '@features/mileage/TripAddressComboBox/TripAddressComboBox'
 import { useTripForm } from '@features/mileage/TripForm/useTripForm'
 import { TripPurposeComboBox } from '@features/mileage/TripPurposeComboBox/TripPurposeComboBox'
