+++ src/components/features/mileage/VehicleForm/vehicleForm.scss
@@ -13,13 +13,14 @@
     padding-block-start: 1rem;
   }
 
+  --form-field-label-inline-size: 10rem;
+
   .Layer__VehicleForm__Field__MakeAndModel,
   .Layer__VehicleForm__Field__Year,
   .Layer__VehicleForm__Field__LicensePlate,
   .Layer__VehicleForm__Field__Vin,
   .Layer__VehicleForm__Field__Description,
   .Layer__VehicleForm__Field__IsPrimary {
-    grid-template-columns: 10rem auto;
     width: 100%;
   }
 }
