+++ src/components/features/timeTracking/TimeEntryForm/timeEntryForm.scss
@@ -10,12 +10,13 @@
     margin-block-start: calc(-1 * var(--spacing-md));
   }
 
+  --form-field-label-inline-size: 10rem;
+
   .Layer__TimeEntryForm__Field__EntryDate,
   .Layer__TimeEntryForm__Field__Duration,
   .Layer__TimeEntryForm__Field__Service,
   .Layer__TimeEntryForm__Field__Customer,
   .Layer__TimeEntryForm__Field__Memo {
-    grid-template-columns: 10rem auto;
     width: 100%;
   }
