+++ src/components/ui/Checkbox/checkbox.scss
@@ -64,6 +64,14 @@
     background-color: var(--color-info-error);
   }
 
+  &[data-invalid] [slot='checkbox'] {
+    border-color: var(--color-danger);
+  }
+
+  &:is([data-selected], [data-indeterminate])[data-invalid] [slot='checkbox'] {
+    background-color: var(--color-danger);
+  }
+
   &[data-variant='round'] [slot='checkbox'] {
     border-radius: 50%;
     border: 0.5px solid var(--color-base-400);
